Added simple TOS logic
