--- app/src/main/java/com/face_location/facelocation/ForTests.java
@@ -1,13 +1,0 @@
-package com.face_location.facelocation;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class ForTests extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_for_tests);
-    }
-}
