+++ app/src/main/res/layout/app_bar_menu.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.face_location.facelocation.MenuActivity">
+    tools:context="com.face_location.facelocation.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
