+++ app/src/main/res/layout/content_menu.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.face_location.facelocation.MenuActivity"
+    tools:context="com.face_location.facelocation.MainActivity"
     tools:showIn="@layout/app_bar_menu">
 
 </android.support.constraint.ConstraintLayout>
