+++ app/src/main/res/layout/map.xml
@@ -5,5 +5,5 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.face_location.facelocation.MapsActivity" />
+    tools:context="com.face_location.facelocation.Map" />
