+++ app/src/main/res/values/strings.xml
@@ -7,8 +7,8 @@
     <string name="app_name_uppercase">FACELOCATION</string>
     <string name="tos_cancel">НЕ ПОГОДЖУЮСЬ</string>
     <string name="tos_accept">ПОГОДЖУЮСЬ</string>
-    <string name="title_activity_maps">Map</string>
-    <string name="title_activity_menu">MenuActivity</string>
+    <string name="title_activity_map">Map</string>
+    <string name="title_activity_main">MainActivity</string>
 
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
